refactor(api): extract admin from client.ts
